lost_animals/requirements.txt
altgraph==0.16.1
asn1crypto==0.24.0
attrs==18.2.0
auto-py-to-exe==2.4.3
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
bottle==0.12.16
bottle-websocket==0.2.9
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
clone==0.0
colorama==0.3.9
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.0
Django==2.1.4
django-filter==2.1.0
django-phonenumber-field==2.3.1
eel==0.9.10
future==0.17.1
gevent-websocket==0.10.1
greenlet==0.4.15
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
joblib==0.12.5
kiwisolver==1.0.1
lxml==4.2.5
macholib==1.11
matplotlib==2.2.3
numpy==1.15.1
pandas==0.23.4
parsel==1.5.1
parso==0.3.1
pefile==2018.8.8
pickleshare==0.7.4
Pillow==6.0.0
prompt-toolkit==1.0.15
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pyfiglet==0.7.5
Pygments==2.2.0
PyHamcrest==1.9.0
pyinstaller==3.4
pyOpenSSL==18.0.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
pywin32-ctypes==0.2.0
queuelib==1.5.0
requests==2.19.1
scikit-learn==0.19.2
scipy==1.1.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
traitlets==4.3.2
urllib3==1.23
virtualenv==16.1.0
w3lib==1.19.0
wcwidth==0.1.7
whichcraft==0.5.2
zope.interface==4.6.0
